integrated new plugins to architecture. renamed title-context-plugin to similar-title-plugin
