+++ config/index.js
@@ -1,6 +1,7 @@
 const ScoreManager = require('../lib/score-manager')
 const aggregator = require('../lib/score-aggregator')
-const sameTitlePlugin = require('../lib/plugins/same-title-plugin')
+const closeTimePlugin = require('../lib/plugins/close-time-plugin')
+const similarTitlePlugin = require('../lib/plugins/similar-title-plugin')
 
 /**
  * Returns a new ScoreManager where the default plugins
@@ -8,7 +9,8 @@ const sameTitlePlugin = require('../lib/plugins/same-title-plugin')
  */
 function makeScoreManager () {
   let manager = new ScoreManager(new aggregator.Mean())
-  manager.registerPlugin('same-title', sameTitlePlugin)
+  manager.registerPlugin('similar-title', similarTitlePlugin)
+  manager.registerPlugin('close-time', closeTimePlugin)
   return manager
 }
