+++ lib/plugins/similar-title-plugin.js
